아이템 아이콘 위치 초기화 코드 추가 및 화면 밖으로 사라질 때 밑으로 내림(기존에는 -5500 위치에 넣어서 문제 발생 소지 있었음)

+++ oldboy/oldboy/RMairTomo.cpp
@@ -78,7 +78,7 @@ void CRMairTomo::SetRandomJudge( UINT NextNoteTime, WidgetType NextNoteType , UI
 		m_KeyInputReadyTime = NextNoteTime + CRMrandomGenerator::GetInstance()->GetRandom(2260 - m_Mental, 2260);
 	}
 
-	printConsole("Mental : %d %d \n", m_Mental, m_KeyInputReadyTime - NextNoteTime);
+	printConsole("Mental : %d %d \n", m_Mental, static_cast<int>(m_KeyInputReadyTime - NextNoteTime));
 
 	m_InputReadyList.push_back(m_KeyInputReadyTime);
 	m_InputReadyType.push_back(NextNoteType);
